requirements.txt
fastapi
uvicorn
pydantic
langchain-groq
python-dotenv
chromadb
beautifulsoup4
playwright
langchain-community
langchain-chroma
langchain-huggingface
huggingface_hub
